feat(FOROME-894): text editor, saving dataset modals fixes
