+++ src/pages/filter/dtree/components/control-panel/text-editor-button.tsx
@@ -1,16 +1,27 @@
-import { observer } from 'mobx-react-lite'
-
+import { useModal } from '@core/hooks/use-modal'
 import { t } from '@i18n'
 import { Button } from '@ui/button'
 import { DecisionTreesMenuDataCy } from '@components/data-testid/decision-tree-menu.cy'
-import modalsVisibilityStore from '../modals/modals-visibility-store'
+import { ModalTextEditor } from '../modals/components/modal-text-editor'
 
-export const TextEditorButton = observer(() => (
-  <Button
-    text={t('dtree.textEditor')}
-    size="md"
-    variant="secondary-dark"
-    onClick={modalsVisibilityStore.openModalTextEditor}
-    dataTestId={DecisionTreesMenuDataCy.textEditor}
-  />
-))
+export const TextEditorButton = () => {
+  const [textEditorDialog, openTextEditorDialog, closeTextEditorDialog] =
+    useModal()
+  const { isOpen } = textEditorDialog
+
+  return (
+    <>
+      <Button
+        text={t('dtree.textEditor')}
+        size="md"
+        variant="secondary-dark"
+        onClick={openTextEditorDialog}
+        dataTestId={DecisionTreesMenuDataCy.textEditor}
+      />
+
+      {isOpen && (
+        <ModalTextEditor closeModal={closeTextEditorDialog} isOpen={isOpen} />
+      )}
+    </>
+  )
+}
