+++ src/pages/filter/dtree/components/modals/components/modal-text-editor.tsx
@@ -14,7 +14,7 @@ import { DecisionTreeModalDataCy } from '@components/data-testid/decision-tree-m
 import Editor from '@monaco-editor/react'
 import { IDtreeCheck } from '@service-providers/decision-trees/decision-trees.interface'
 import { getMessageFromError } from '@utils/http/getMessageFromError'
-import modalsVisibilityStore from '../modals-visibility-store'
+import { ICommonModalProps } from '../modals.interface'
 
 const TEXT_EDITOR_THEME = 'textEditorTheme'
 
@@ -45,178 +45,180 @@ const hasError = function (error: typeof emptyError) {
   return error.error?.length > 0
 }
 
-export const ModalTextEditor = observer((): ReactElement => {
-  const params = useParams()
-  const [checked, setChecked] = useState(true)
+export const ModalTextEditor = observer(
+  ({ isOpen, closeModal }: ICommonModalProps): ReactElement => {
+    const params = useParams()
+    const [checked, setChecked] = useState(true)
 
-  const [code, setCode] = useState(dtreeStore.dtreeCode)
+    const [code, setCode] = useState(dtreeStore.dtreeCode)
 
-  useEffect(() => {
-    if (dtreeStore.localDtreeCode) {
-      dtreeStore.setNextDtreeCode(dtreeStore.localDtreeCode)
-      setCode(dtreeStore.localDtreeCode)
-      dtreeStore.resetLocalDtreeCode()
-    } else {
-      dtreeStore.setStartDtreeCode()
-    }
-  }, [])
+    useEffect(() => {
+      if (dtreeStore.localDtreeCode) {
+        dtreeStore.setNextDtreeCode(dtreeStore.localDtreeCode)
+        setCode(dtreeStore.localDtreeCode)
+        dtreeStore.resetLocalDtreeCode()
+      } else {
+        dtreeStore.setStartDtreeCode()
+      }
+    }, [])
 
-  const [theme, setTheme] = useState<TEditorTheme>(
-    LocalStoreManager.read<TEditorTheme>(TEXT_EDITOR_THEME) || 'light',
-  )
-
-  const [error, setError] = useState(emptyError)
-
-  const handleDtreeCheckAsync = debounce(async (codeToCheck: string) => {
-    setChecked(false)
-
-    setCode(codeToCheck)
-
-    const response = await fetchDtreeCheckAsync(
-      params.get('ds') || '',
-      codeToCheck,
+    const [theme, setTheme] = useState<TEditorTheme>(
+      LocalStoreManager.read<TEditorTheme>(TEXT_EDITOR_THEME) || 'light',
     )
 
-    setError(
-      response.status === 500
-        ? {
-            error: t('dtree.expressionIsNotCorrect'),
-            line: 0,
-            pos: 0,
-          }
-        : emptyError,
-    )
+    const [error, setError] = useState(emptyError)
 
-    if (response.ok) {
-      const result: IDtreeCheck = await response.json()
+    const handleDtreeCheckAsync = debounce(async (codeToCheck: string) => {
+      setChecked(false)
+
+      setCode(codeToCheck)
+
+      const response = await fetchDtreeCheckAsync(
+        params.get('ds') || '',
+        codeToCheck,
+      )
 
       setError(
-        result.error
+        response.status === 500
           ? {
-              error: result.error,
-              line: result.line as number,
-              pos: result.pos as number,
+              error: t('dtree.expressionIsNotCorrect'),
+              line: 0,
+              pos: 0,
             }
           : emptyError,
       )
-    } else {
-      const errorText = await response.text()
-      const errorNormalized = getMessageFromError(errorText, response.status)
 
-      setError({
-        error: errorNormalized.message,
-        line: 0,
-        pos: 0,
+      if (response.ok) {
+        const result: IDtreeCheck = await response.json()
+
+        setError(
+          result.error
+            ? {
+                error: result.error,
+                line: result.line as number,
+                pos: result.pos as number,
+              }
+            : emptyError,
+        )
+      } else {
+        const errorText = await response.text()
+        const errorNormalized = getMessageFromError(errorText, response.status)
+
+        setError({
+          error: errorNormalized.message,
+          line: 0,
+          pos: 0,
+        })
+      }
+
+      setChecked(true)
+    }, 300)
+
+    const handleDrop = () => {
+      dtreeStore.setNextDtreeCode(dtreeStore.startDtreeCode)
+      handleDtreeCheckAsync(dtreeStore.startDtreeCode)
+
+      setCode(dtreeStore.startDtreeCode)
+    }
+
+    const handleDone = () => {
+      if (code !== dtreeStore.dtreeCode) {
+        dtreeStore.setLocalDtreeCode(code)
+      }
+
+      closeModal()
+    }
+
+    const handleSave = () => {
+      dtreeStore.setNextDtreeCode(dtreeStore.startDtreeCode)
+
+      dtreeStore.fetchDtreeSetAsync({
+        ds: datasetStore.datasetName,
+        code,
+      })
+
+      closeModal()
+    }
+
+    const handleChangeTheme = () => {
+      setTheme(prev => {
+        const themeToChange: TEditorTheme = prev === 'light' ? 'dark' : 'light'
+
+        LocalStoreManager.write(TEXT_EDITOR_THEME, themeToChange)
+
+        return themeToChange
       })
     }
 
-    setChecked(true)
-  }, 300)
-
-  const handleDrop = () => {
-    dtreeStore.setNextDtreeCode(dtreeStore.startDtreeCode)
-    handleDtreeCheckAsync(dtreeStore.startDtreeCode)
-
-    setCode(dtreeStore.startDtreeCode)
-  }
-
-  const handleDone = () => {
-    if (code !== dtreeStore.dtreeCode) {
-      dtreeStore.setLocalDtreeCode(code)
-    }
-
-    modalsVisibilityStore.closeModalTextEditor()
-  }
-
-  const handleSave = () => {
-    dtreeStore.setNextDtreeCode(dtreeStore.startDtreeCode)
-
-    dtreeStore.fetchDtreeSetAsync({
-      ds: datasetStore.datasetName,
-      code,
-    })
-
-    modalsVisibilityStore.closeModalTextEditor()
-  }
-
-  const handleChangeTheme = () => {
-    setTheme(prev => {
-      const themeToChange: TEditorTheme = prev === 'light' ? 'dark' : 'light'
-
-      LocalStoreManager.write(TEXT_EDITOR_THEME, themeToChange)
-
-      return themeToChange
-    })
-  }
-
-  const Controls = (): ReactElement => (
-    <>
-      <Button
-        text="Drop changes"
-        size="md"
-        onClick={handleDrop}
-        variant={theme === 'light' ? 'secondary' : 'secondary-dark'}
-      />
-
-      <Button
-        text="Done"
-        size="md"
-        disabled={!checked || hasError(error)}
-        onClick={handleDone}
-        variant={theme === 'light' ? 'secondary' : 'secondary-dark'}
-      />
-
-      <Button
-        text="Save"
-        size="md"
-        disabled={!checked || hasError(error)}
-        onClick={handleSave}
-        variant={theme === 'light' ? 'primary' : 'primary-dark'}
-      />
-    </>
-  )
-
-  return (
-    <Dialog
-      isOpen={modalsVisibilityStore.isModalTextEditorVisible}
-      onClose={() => modalsVisibilityStore.closeModalTextEditor()}
-      title={t('dtree.editCurrentDecisionTreeCode')}
-      isApplyDisabled={!checked || hasError(error)}
-      width="xl"
-      data-testid={DecisionTreeModalDataCy.modalHeader}
-      handleChangeTheme={handleChangeTheme}
-      theme={theme}
-      style={{ top: '50%' }}
-      actions={<Controls />}
-    >
-      <div className="flex items-center mt-1">
-        {hasError(error) && (
-          <div className="text-red-secondary bg-yellow-bright">
-            {error.line !== 0 && error.pos !== 0
-              ? `At line ${error.line} pos ${error.pos}: ${error.error}`
-              : `${error.error}`}
-          </div>
-        )}
-      </div>
-
-      <div className="flex-1 mt-3 overflow-hidden">
-        <Editor
-          height="65vh"
-          defaultLanguage="python"
-          value={code}
-          options={{
-            cursorSmoothCaretAnimation: true,
-            minimap: {
-              showSlider: 'always',
-            },
-          }}
-          theme={theme === 'light' ? 'vs-white' : 'vs-dark'}
-          onChange={(e: any) => {
-            handleDtreeCheckAsync(e)
-          }}
-          className="border border-grey-blue"
+    const Controls = (): ReactElement => (
+      <>
+        <Button
+          text="Drop changes"
+          size="md"
+          onClick={handleDrop}
+          variant={theme === 'light' ? 'secondary' : 'secondary-dark'}
         />
-      </div>
-    </Dialog>
-  )
-})
+
+        <Button
+          text="Done"
+          size="md"
+          disabled={!checked || hasError(error)}
+          onClick={handleDone}
+          variant={theme === 'light' ? 'secondary' : 'secondary-dark'}
+        />
+
+        <Button
+          text="Save"
+          size="md"
+          disabled={!checked || hasError(error)}
+          onClick={handleSave}
+          variant={theme === 'light' ? 'primary' : 'primary-dark'}
+        />
+      </>
+    )
+
+    return (
+      <Dialog
+        isOpen={isOpen}
+        onClose={closeModal}
+        title={t('dtree.editCurrentDecisionTreeCode')}
+        isApplyDisabled={!checked || hasError(error)}
+        width="xl"
+        data-testid={DecisionTreeModalDataCy.modalHeader}
+        handleChangeTheme={handleChangeTheme}
+        theme={theme}
+        style={{ top: '50%' }}
+        actions={<Controls />}
+      >
+        <div className="flex items-center">
+          {hasError(error) && (
+            <div className="text-red-secondary bg-yellow-bright">
+              {error.line !== 0 && error.pos !== 0
+                ? `At line ${error.line} pos ${error.pos}: ${error.error}`
+                : `${error.error}`}
+            </div>
+          )}
+        </div>
+
+        <div className="mt-1 overflow-hidden">
+          <Editor
+            height="65vh"
+            defaultLanguage="python"
+            value={code}
+            options={{
+              cursorSmoothCaretAnimation: true,
+              minimap: {
+                showSlider: 'always',
+              },
+            }}
+            theme={theme === 'light' ? 'vs-white' : 'vs-dark'}
+            onChange={(e: any) => {
+              handleDtreeCheckAsync(e)
+            }}
+            className="border border-grey-blue"
+          />
+        </div>
+      </Dialog>
+    )
+  },
+)
