+++ src/pages/filter/dtree/components/modals/modals-container.tsx
@@ -12,9 +12,7 @@ import { ModalEnum } from './components/modal-enum/modal-enum'
 import { ModalGeneRegion } from './components/modal-gene-region/modal-gene-region'
 import { ModalInheritanceMode } from './components/modal-inheritance-mode/modal-inheritance-mode'
 import { ModalNumeric } from './components/modal-numeric'
-import { ModalSaveDataset } from './components/modal-save-dataset'
 import { ModalSelectAttribute } from './components/modal-select-attribute'
-import { ModalTextEditor } from './components/modal-text-editor'
 import { ModalViewVariants } from './components/modal-view-variants'
 import modalsVisibilityStore from './modals-visibility-store'
 
@@ -46,10 +44,7 @@ export const ModalsContainer = observer(
       {modalsVisibilityStore.isModalGeneRegionVisible && <ModalGeneRegion />}
 
       {dtreeStore.isModalViewVariantsVisible && <ModalViewVariants />}
-      {modalsVisibilityStore.isModalTextEditorVisible && <ModalTextEditor />}
-      {dtreeStore.isModalSaveDatasetVisible && <ModalSaveDataset />}
 
-      {handleDatasetStore.isExportModalShown}
       {handleDatasetStore.isExportModalShown && <ExportModal />}
       {handleDatasetStore.isImportModalShown && <ImportModal />}
     </Fragment>
