+++ src/pages/filter/dtree/components/modals/modals.interface.tsx
@@ -1,0 +1,4 @@
+export interface ICommonModalProps {
+  closeModal: () => void
+  isOpen: boolean
+}
