+++ src/store/dtree/dtree.store.ts
@@ -105,7 +105,6 @@ export class DtreeStore {
   stepData: IStepData[] = []
   stepAmout = 0
 
-  isModalSaveDatasetVisible = false
   isModalViewVariantsVisible = false
   tableModalIndexNumber: null | number = null
 
@@ -512,14 +511,6 @@ export class DtreeStore {
   }
 
   // 3.4 Common UI/UX modals
-
-  openModalSaveDataset = () => {
-    this.isModalSaveDatasetVisible = true
-  }
-
-  closeModalSaveDataset() {
-    this.isModalSaveDatasetVisible = false
-  }
 
   openModalViewVariants(index?: number) {
     this.isModalViewVariantsVisible = true
